Switch between search clients using an env var
